+++ pkg/services/sqlstore/annotation_test.go
@@ -42,6 +42,7 @@ func TestAnnotations(t *testing.T) {
 				UserId:      1,
 				DashboardId: 1,
 				Text:        "hello",
+				Type:        "alert",
 				Epoch:       10,
 				Tags:        []string{"outage", "error", "type:outage", "server:server-1"},
 			}
@@ -85,6 +86,19 @@ func TestAnnotations(t *testing.T) {
 					From:        1,
 					To:          15,
 					Tags:        []string{"asd"},
+				})
+
+				So(err, ShouldBeNil)
+				So(items, ShouldHaveLength, 0)
+			})
+
+			Convey("Should not find one when type filter does not match", func() {
+				items, err := repo.Find(&annotations.ItemQuery{
+					OrgId:       1,
+					DashboardId: 1,
+					From:        1,
+					To:          15,
+					Type:        "alert",
 				})
 
 				So(err, ShouldBeNil)